+++ sdk/lib/_internal/vm/lib/async_patch.dart
@@ -35,7 +35,8 @@ class _AsyncAwaitCompleter<T> implements Completer<T> {
     }
   }
 
-  void completeError(e, [st]) {
+  void completeError(Object e, [StackTrace st]) {
+    st ??= AsyncError.defaultStackTrace(e);
     if (isSync) {
       _future._completeError(e, st);
     } else {
@@ -79,7 +80,7 @@ Function _asyncThenWrapperHelper(continuation) {
 // parameter to the continuation.  See vm/ast_transformer.cc for usage.
 Function _asyncErrorWrapperHelper(continuation) {
   // See comments of `_asyncThenWrapperHelper`.
-  var errorCallback = (e, s) => continuation(null, e, s);
+  void errorCallback(Object e, StackTrace s) => continuation(null, e, s);
   if (Zone.current == Zone.root) return errorCallback;
   return Zone.current.registerBinaryCallback(errorCallback);
 }