+++ sdk/lib/async/broadcast_stream_controller.dart
@@ -260,6 +260,7 @@ abstract class _BroadcastStreamController<T>
       error = _nonNullError(replacement.error);
       stackTrace = replacement.stackTrace;
     }
+    stackTrace ??= AsyncError.defaultStackTrace(error);
     _sendError(error, stackTrace);
   }
 
@@ -481,6 +482,7 @@ class _AsBroadcastStreamController<T> extends _SyncBroadcastStreamController<T>
 
   void addError(Object error, [StackTrace stackTrace]) {
     ArgumentError.checkNotNull(error, "error");
+    stackTrace ??= AsyncError.defaultStackTrace(error);
     if (!isClosed && _isFiring) {
       _addPendingEvent(new _DelayedError(error, stackTrace));
       return;